requirements.txt
alembic==1.13.1
Faker==23.1.0
Flask==3.0.2
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
flask_login==0.6.3
SQLAlchemy==2.0.25
WTForms==3.1.2
email_validator==2.1.0.post1
djlint==1.34.1
python-dotenv==1.0.1
psycopg2==2.9.9
flask_mail==0.9.1
pyjwt==2.8.0